refactorization of the code. two UI consume the logic
